fix(illegal): webpack is doing something weird
Intermittently, the minified file is throwing this error:

> Uncaught SyntaxError: Unexpected token ILLEGAL

It doesn't make any sense, because simply opening the file and
saving it again causes the error to go away — for a while.

So I'm thinking maybe minification is the issue. Failing that,
I'll see what happens if we bundle vendor scripts separately.